add the magic header # encoding: utf-8  to a few files, plus refactoring the hot threads code to be in single location
