+++ logstash-core/lib/logstash/api/commands/base.rb
@@ -1,3 +1,5 @@
+# encoding: utf-8
+
 module LogStash
   module Api
     module Commands
