+++ logstash-core/lib/logstash/api/commands/default_metadata.rb
@@ -1,3 +1,5 @@
+# encoding: utf-8
+
 require "logstash/api/commands/base"
 
 module LogStash
